Dzień 3.0 (SceneBuilder,Java,Walki)

+++ src/main/java/com/example/gra/Geralt.java
@@ -8,7 +8,7 @@ import java.util.Random;
 public class Geralt {
     public static double power = 10;
     public static int money;
-    public static double CurrentHP = 100;
+    public static int CurrentHP = 100;
     public static int MaxHP = 100;
 
     public static double onNormalAttack(ActionEvent actionEvent) throws IOException {
